+++ src/components/UniqueId.js
@@ -26,31 +26,95 @@ export function unregisterUniqueId(uniqueId) {
 
 const html4Id = /^[A-Za-z]+[\w\-:.]*$/
 
-export function withUniqueId(Component, identifier = 'maria-uid') {
+export function withUniqueId(options) {
+    const { identifier = 'maria-uid', includeGenerator = false } = options || {}
+
     // eslint-disable-next-line no-undef
     if (process.env.NODE_ENV !== 'production') {
-        if (!html4Id.test(identifier)) {
+        if (identifier == null || !html4Id.test(identifier)) {
             throw new Error('[withUniqueId] identifier must be a HTML4 compatible id string starting with ASCII letter')
         }
     }
 
-    class UniqueId extends React.PureComponent {
-        static propTypes = {
-            forwardedRef: PropTypes.func,
+    return function(Component) {
+        class UniqueId extends React.PureComponent {
+            static propTypes = {
+                forwardedRef: PropTypes.any,
+                uniqueId: PropTypes.string,
+            }
+
+            nextIds = {}
+            state = {
+                uniqueId: registerUniqueId(this.props.uniqueId != null ? this.props.uniqueId : identifier),
+                uniqueIdProps: this.props.uniqueId,
+            }
+
+            static getDerivedStateFromProps(props, state) {
+                if (state.uniqueIdProps === props.uniqueId) {
+                    return null
+                }
+
+                unregisterUniqueId(state.uniqueId)
+
+                return {
+                    uniqueId: registerUniqueId(props.uniqueId != null ? props.uniqueId : identifier),
+                    uniqueIdProps: props.uniqueId,
+                }
+            }
+
+            componentWillUnmount() {
+                unregisterUniqueId(this.state.uniqueId)
+            }
+
+            lastUniqueId = id => {
+                const key = typeof id === 'string' ? id : ''
+                const array = this.nextIds[key]
+
+                return (Array.isArray(array) && array[array.length - 1]) || null
+            }
+
+            nextUniqueId = id => {
+                const key = typeof id === 'string' ? id : ''
+                const uniqueId = registerUniqueId(`${this.state.uniqueId}.${key}`)
+
+                if (!Array.isArray(this.nextIds[key])) {
+                    this.nextIds[key] = []
+                }
+
+                this.nextIds[key].push(uniqueId)
+                return uniqueId
+            }
+
+            generator = Object.defineProperty(Object.defineProperty({}, 'last', { value: this.lastUniqueId }), 'next', {
+                value: this.nextUniqueId,
+            })
+
+            uniqueIdGen = () => {
+                const keys = Object.keys(this.nextIds)
+
+                for (let j = 0; j < keys.length; j++) {
+                    const key = keys[j]
+
+                    for (let i = 0; i < this.nextIds[key].length; i++) {
+                        unregisterUniqueId(this.nextIds[key][i])
+                    }
+
+                    this.nextIds[key].length = 0
+                }
+
+                return this.generator
+            }
+
+            render() {
+                const { forwardedRef, ...props } = this.props
+                const { uniqueId } = this.state
+
+                const uniqueIdGen = includeGenerator ? this.uniqueIdGen : null
+
+                return <Component {...props} ref={forwardedRef} uniqueId={uniqueId} uniqueIdGen={uniqueIdGen} />
+            }
         }
 
-        uniqueId = registerUniqueId(identifier)
-
-        componentWillUnmount() {
-            unregisterUniqueId(this.uniqueId)
-        }
-
-        render() {
-            const { forwardedRef, ...props } = this.props
-
-            return <Component {...props} ref={forwardedRef} uniqueId={this.uniqueId} />
-        }
+        return React.forwardRef((props, ref) => <UniqueId {...props} forwardedRef={ref} />)
     }
-
-    return React.forwardRef((props, ref) => <UniqueId {...props} forwardedRef={ref} />)
 }