+++ src/index.js
@@ -1,4 +1,4 @@
-export { default as Overlay } from './components/Overlay'
-export { default as withOverlay } from './components/OverlayContext'
-export { default as OverlayProvider } from './components/OverlayProvider'
+export { default as Modal } from './components/Modal'
+export { default as ModalOverlay } from './components/ModalOverlay'
+export { default as withOverlay } from './components/Overlay'
 export { clearUniqueIds, registerUniqueId, unregisterUniqueId, withUniqueId } from './components/UniqueId'